- clean dor struct - only keeping final files
