+++ .gitignore
@@ -1,3 +1,3 @@
-MyWork/venv/
-MyWork/__pycache__/
-MyWork/app/__pycache__/
+.venv
+__pycache__
+.vscode
