--- MyWork/app/__init__.py
@@ -1,62 +1,0 @@
-# Initialize Flask Application 
-
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_socketio import SocketIO
-from flask_migrate import Migrate
-
-db = SQLAlchemy()
-socketio = SocketIO()
-migrate = Migrate()
-
-def create_app():
-    app = Flask(__name__)
-    app.config.from_object('app.config.Config')
-
-    db.init_app(app)
-    migrate.init_app(app, db)
-    socketio.init_app(app)
-
-    with app.app_context():
-        from app import routes  # Import routes inside the app context
-        db.create_all()
-
-    return app
-
-app = create_app()
-
-if __name__ == '__main__':
-    socketio.run(app, host='0.0.0.0', port=5000)
-
-
-
-
-# # Initialize Flask Application 
-
-# from flask import Flask
-# from flask_sqlalchemy import SQLAlchemy
-# from flask_socketio import SocketIO
-# from flask_migrate import Migrate
-
-# db = SQLAlchemy()
-# socketio = SocketIO()
-# migrate = Migrate()
-
-# def create_app():
-#     app = Flask(__name__)
-#     app.config.from_object('app.config.Config')
-
-#     db.init_app(app)
-#     migrate.init_app(app, db)
-#     socketio.init_app(app)
-
-#     with app.app_context():
-#         from app import routes  # Import routes inside the app context
-#         db.create_all()
-
-#     return app
-
-# app = create_app()
-
-# if __name__ == '__main__':
-#     socketio.run(app, host='0.0.0.0', port=5000)
