--- MyWork/app/config.py
@@ -1,32 +1,0 @@
-import os
-
-basedir = os.path.abspath(os.path.dirname(__file__))
-
-class Config:
-    SECRET_KEY = os.getenv('SECRET_KEY', 'default_secret_key')
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, 'app.db')
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-    SQLALCHEMY_ENGINE_OPTIONS = {
-        'pool_pre_ping': True,
-        'pool_recycle': 300,
-        'pool_size': 10,
-        'max_overflow': 20
-    }
-
-
-
-# Following is for PostgresSql
-
-# # Configuration Settings
-# import os
-
-# class Config:
-#     SECRET_KEY = os.getenv('SECRET_KEY', 'default_secret_key')
-#     SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL', 'postgresql+psycopg2://postgres:123@localhost/health_monitor')
-#     SQLALCHEMY_TRACK_MODIFICATIONS = False
-#     SQLALCHEMY_ENGINE_OPTIONS = {
-#         'pool_pre_ping': True,
-#         'pool_recycle': 300,
-#         'pool_size': 10,
-#         'max_overflow': 20
-#     }
