+++ README.md
@@ -2,35 +2,40 @@
 
 ## Section 1: Installation (Only One Time Setup)
 
-1. Enter the root folder
+Note: Only applicable for Windows OS 
+
+1. Enter the root folder ./K270
 
 2. Create a virtual environment:
     ```bash
-    python -m venv venv
+    python -m venv .venv
     ``` 
 
 2. Activate the venv:
 	```bash
-    .\venv\Scripts\activate
+    .\.venv\Scripts\activate
     ```
 
+3. Updgrade pip:
+	```bash
+    python -m pip install --upgrade pip
+    ```
 
-3. Install the required packages:
+4. Install the required packages:
     ```bash
     pip install -r requirements.txt
     ```
 
 ## Section 2: Running
 
-1. Navigate to the `MyWork` folder:
-   
+Note: Only applicable for Windows OS 
 
-2. Initiate the virtual environment (for Windows):
+1. Initiate the virtual environment:
     ```bash
-    .\venv\Scripts\activate
+    .\.venv\Scripts\activate
     ```
 
-3. Use the following script for hosting on a local machine:
+2. Use the following script for hosting on a local machine:
     ```bash
 	set FLASK_APP=app
     set FLASK_ENV=development
@@ -38,7 +43,7 @@
     ```
 	This will show IP similar to: http://127.0.0.1:5000. Use this in the same device's browser of your choice.
 
-4. Use the following script for hosting on a LAN:
+3. Use the following script for hosting on a LAN:
     ```bash
 	set FLASK_APP=app
     set FLASK_ENV=development
@@ -48,16 +53,16 @@
 	
 ## Section 3: Testing (While the server is running in one terminal, open another cmd)
 
-1. Navigate to MyWork\ 
+Note: Only applicable for Windows OS 
 
-2. Activate Virtual environment again this new terminal:
+1. Activate Virtual environment again this new terminal:
 	
 	```bash
     .\venv\Scripts\activate
     ```
-3. Navigate to MyWork\tests\ 
+2. Navigate to .\tests\ 
 
-4. Run python script, and monitor the browser on your screen. Following script will send 10 dummy sensor data to populate the database, and hence updates the UI in real-time:
+3. Run python script, and monitor the browser on your screen. Following script will send 10 dummy sensor data to populate the database, and hence updates the UI in real-time:
 	```bash
 	 python send_data_OneSecond.py
     ```
