+++ app/static/css/style.css
@@ -9,7 +9,8 @@ body {
     align-items: center;
     height: 100vh;
     overflow-y: auto;
-    overflow-x: auto; /* Prevent body scroll */
+    overflow-x: auto;
+    /* Prevent body scroll */
 }
 
 .container {
@@ -39,9 +40,9 @@ body {
 }
 
 .content {
-    padding-top: 100px; /* Adjust based on title height */
+    padding-top: 100px;
     width: 100%;
-    height: calc(100vh - 70px); /* Adjust based on title height */
+    height: calc(100vh - 70px);
     overflow-y: auto;
     overflow-x: auto;
 }
@@ -149,6 +150,11 @@ h2 {
 }
 
 @keyframes fadeIn {
-    from { opacity: 0; }
-    to { opacity: 1; }
+    from {
+        opacity: 0;
+    }
+
+    to {
+        opacity: 1;
+    }
 }
