+++ app/static/js/app.js
@@ -214,7 +214,7 @@ document.addEventListener("DOMContentLoaded", function() {
         if ('Notification' in window && Notification.permission === 'granted') {
             new Notification("Fall Detected!", {
                 body: message,
-                icon: '/static/img/alert-icon.png' // Add an icon if desired
+                icon: '/static/img/alert-icon.png'
             });
         }
